NCalc is now linked as as a package reference.
There is no need to modifiy NCalc code, since it can already handle operations with parameter vectors. It even requests the enumerator from the parameter object, so we only need to implement an interator this returns random items to the expression evaluator. 

NCalc internally does not implement a super-efficient vector operation, but it should be more than enough for this kind of models. 

If more performant models need to be created, the operations should be definitely implemented using the Distribution.Apply methods instead of the expression evaluation of NCalc.